band grouping - extract inner classes to outer, refactor mark old as deprecated
