+++ opttbx-sentinel3-reader/src/test/java/eu/esa/opt/dataio/s3/slstr/MetaDataProviderTest.java
@@ -1,9 +1,8 @@
 package eu.esa.opt.dataio.s3.slstr;
 
 import com.bc.ceres.annotation.STTM;
-import eu.esa.opt.snap.core.datamodel.band.DataPoint;
-import eu.esa.opt.snap.core.datamodel.band.SparseDataBand;
-import eu.esa.opt.snap.core.datamodel.band.SparseDataProvider;
+import eu.esa.snap.core.datamodel.band.DataPoint;
+import eu.esa.snap.core.datamodel.band.SparseDataBand;
 import org.esa.snap.core.datamodel.MetadataAttribute;
 import org.esa.snap.core.datamodel.MetadataElement;
 import org.esa.snap.core.datamodel.Product;
